Add files via upload
This repository contains code and results derived from the article focusing on the closed-form solution for linear and nonlinear free vibrations of composite and fiber metal laminated (FML) rectangular plates. The project aims to reproduce and extend the results presented in the article, which provides analytical solutions for the free vibration behavior of composite plates and FML plates under various conditions.